Refactoring tools to concrete package per tool
