+++ src/tools/Highlighter/index.js
