+++ src/tools/Rectangle/rectangle-label-object.js
