+++ src/tools/index.js
@@ -1,0 +1,31 @@
+import Arrow from './Arrow'
+import Circle from './Circle'
+import Highlighter from './Highlighter'
+import Line from './Line'
+import Pan from './Pan'
+import Pencil from './Pencil'
+import Rectangle from './Rectangle'
+import RectangleLabel from './Rectangle/rectangle-label'
+import Select from './Select'
+
+export {Arrow}
+export {Circle}
+export {Highlighter}
+export {Line}
+export {Pan}
+export {Pencil}
+export {Rectangle}
+export {RectangleLabel}
+export {Select}
+
+export default {
+    Arrow,
+    Circle,
+    Highlighter,
+    Line,
+    Pan,
+    Pencil,
+    Rectangle,
+    RectangleLabel,
+    Select
+}
